refactor vector store types
